refactor(pagination): extract shared FetchNextPageInput type
